Create array_maker.py
This is how I was making the arrays, with a database, that went into the graphs. The plotting function is in a separate file. A quirk I noticed was that you could make edits to the plotting file and see those changes in the ipython kernel but changes to the imported  array_maker.py file are not seen until a new kernel was loaded.