Create map_displacements.py
My attempt to interpolate and plot the data for sites in a given region (ex. So Cal). Only returns plots (east, north, and up) for one time.
Unfortunately the code fails when it tries to plot because for some reason it identifies only two sites in the GPS data that are in the given region.